+++ Sources/MapboxNavigationCore/Map/Other/MapboxMap+Async.swift
@@ -1,5 +1,5 @@
 import _MapboxNavigationHelpers
-import MapboxMaps
+@_spi(Experimental) import MapboxMaps
 
 extension MapboxMap {
     @MainActor
@@ -31,6 +31,25 @@ extension MapboxMap {
         return try await withTaskCancellationHandler {
             try await withCheckedThrowingContinuation { continuation in
                 let cancellable = queryRenderedFeatures(with: rect, options: options) { result in
+                    continuation.resume(with: result)
+                }
+                state.activate(with: .init(cancellable))
+            }
+        } onCancel: {
+            state.cancel()
+        }
+    }
+
+    @MainActor
+    func queryRenderedFeatures<F: FeaturesetFeatureType>(
+        with rect: CGRect,
+        featureset: FeaturesetDescriptor<F>
+    ) async throws -> [F] {
+        let state: CancellableAsyncState<AnyMapboxMapsCancelable> = .init()
+
+        return try await withTaskCancellationHandler {
+            try await withCheckedThrowingContinuation { continuation in
+                let cancellable = queryRenderedFeatures(with: rect, featureset: featureset) { result in
                     continuation.resume(with: result)
                 }
                 state.activate(with: .init(cancellable))